refactor comment_count to be an INT

+++ models/articles.models.js
@@ -20,7 +20,7 @@ exports.selectAllArticles = (topic, sortBy = "created_at", order = "DESC") => {
   }
 
   let queryString = ` 
-      SELECT articles.author, articles.title, articles.article_id, articles.topic, articles.created_at, articles.votes, articles.article_img_url, COUNT (comments.article_id) AS comment_count
+      SELECT articles.author, articles.title, articles.article_id, articles.topic, articles.created_at, articles.votes, articles.article_img_url, COUNT (comments.article_id)::INT AS comment_count
       FROM articles
       FULL OUTER JOIN comments ON articles.article_id = comments.article_id`;
